Compose StructuredTraceGraph from Entities and events graph
